Fixed missing headers for subtitle download. Refactored functions into utils module
